backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
python-dotenv==1.1.0
httpx==0.28.1
pydantic==2.11.4
anthropic==0.52.0
transformers==4.51.3
torch==2.7.0
